+++ src/versioningit/core.py
@@ -234,7 +234,11 @@ class Versioningit:
     def do_onbuild(
         self, build_dir: Union[str, Path], is_source: bool, version: str
     ) -> None:
-        """Run the ``onbuild`` step"""
+        """
+        .. versionadded:: 1.1.0
+
+        Run the ``onbuild`` step
+        """
         if self.onbuild is not None:
             self.onbuild(build_dir=build_dir, is_source=is_source, version=version)
         else:
@@ -368,5 +372,34 @@ def run_onbuild(
     project_dir: Union[str, Path] = os.curdir,
     config: Optional[dict] = None,
 ) -> None:
+    """
+    .. versionadded:: 1.1.0
+
+    Run the ``onbuild`` step for the given project.  If ``config`` is `None`,
+    then ``project_dir`` must contain a :file:`pyproject.toml` file containing
+    a ``[tool.versioningit]`` table; if it does not, a `NotVersioningitError`
+    is raised.
+
+    If ``config`` is not `None`, then any :file:`pyproject.toml` file in
+    ``project_dir`` will be ignored, and the configuration will be taken from
+    ``config`` instead.  ``config`` must be a `dict` whose structure mirrors
+    the structure of the ``[tool.versioningit]`` table in
+    :file:`pyproject.toml`.
+
+    :param build_dir: The directory containing the in-progress build
+    :param is_source:
+        Set to `True` if building an sdist or other artifact that preserves
+        source paths, `False` if building a wheel or other artifact that uses
+        installation paths
+    :param version: The project's version
+    :raises NotVersioningitError:
+        - if ``config`` is `None` and ``project_dir`` does not contain a
+          :file:`pyproject.toml` file
+        - if the :file:`pyproject.toml` file does not contain a
+          ``[tool.versioningit]`` table
+    :raises ConfigError:
+        if any of the values in ``config`` are not of the correct type
+    :raises MethodError: if a method returns a value of the wrong type
+    """
     vgit = Versioningit.from_project_dir(project_dir, config)
     vgit.do_onbuild(build_dir=build_dir, is_source=is_source, version=version)